requirements.txt
streamlit==1.15.2
requests~=2.28.1
Pillow~=9.2.0
streamlit-lottie==0.0.3